+++ macroCounter/Days.h
@@ -14,29 +14,29 @@ public:
 	void SetDate();
 	void SetDateFile(string date);
 	void SetPFCTotals(vector<Meal> mLog);
-	void SetProteinTotal(int protein);
-	void SetCarbTotal(int carb);
-	void SetFatTotal(int fat);
+	void SetProteinTotal(float protein);
+	void SetCarbTotal(float carb);
+	void SetFatTotal(float fat);
 	void SetCals();
-	void SetCalsTotal(int cals);
+	void SetCalsTotal(float cals);
 	string GetDate();
-	int GetProteinTotal();
-	int GetCarbTotal();
-	int GetFatTotal();
-	int GetProteinCals();
-	int GetCarbCals();
-	int GetFatCals();
-	int GetTotalCals();
+	float GetProteinTotal();
+	float GetCarbTotal();
+	float GetFatTotal();
+	float GetProteinCals();
+	float GetCarbCals();
+	float GetFatCals();
+	float GetTotalCals();
 	~Days();
 private:
 	string dDate;
-	int proteinTotal;
-	int carbTotal;
-	int fatTotal;
-	int proteinCals;
-	int fatCals;
-	int carbCals;
-	int totalCals;
+	float proteinTotal;
+	float carbTotal;
+	float fatTotal;
+	float proteinCals;
+	float fatCals;
+	float carbCals;
+	float totalCals;
 };
 
 #endif
